Make SvnLibraryAttribute internal
To make the reference assembly working, all visible attributes (incl. assembly level) need the ability that they can be instantiated from the reflection type loader. But attributes defined in the same assembly for which the reference assembly will be generated, they can't be instantiated because of the missing implementation. So make these attributes internal that they will be striped of the reference assembly.